pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openai-project-roles"
version = "1.0.0"
description = "Streamlit app to apply default project roles to OpenAI API platform projects"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "wxpython",
    "importlib_resources>=5.12; python_version < '3.9'",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "freezegun>=1.4.0",
]
lint = [
    "ruff>=0.6.0",
]

[project.scripts]
project-roles = "openai_project_roles.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["openai_project_roles*"]

[tool.setuptools.package-data]
openai_project_roles = ["default_project_roles.yaml","openai_project_budgets.yaml","openai_project_usage.csv"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py38"
extend-exclude = ["dist", ".venv", "openai_project_roles.egg-info"]
